Move Rollup configs to JS
