+++ packages/liveblocks-react-tiptap/.eslintrc.js
@@ -92,5 +92,11 @@ module.exports = {
         "@typescript-eslint/no-floating-promises": "off",
       },
     },
+    {
+      files: ["rollup.config.mjs"],
+      env: {
+        node: true,
+      },
+    },
   ],
 };
