+++ packages/liveblocks-react-tiptap/tsconfig.json
@@ -18,5 +18,5 @@
     // TODO: Try to get rid of these overrides
     "noUncheckedIndexedAccess": false // OVERWRITTEN to relax a bit
   },
-  "include": ["**/*.ts", "**/*.tsx"]
+  "include": ["**/*.ts", "**/*.tsx", "rollup.config.mjs"]
 }
