+++ packages/liveblocks-react-ui/tsconfig.json
@@ -19,6 +19,6 @@
     "plugins/**/*",
     "src/**/*",
     "styles/**/*",
-    "rollup.config.ts"
+    "rollup.config.mjs"
   ]
 }
